Material mappings

+++ src/main/java/org/zibble/inventoryframework/protocol/Material.java
@@ -7,13 +7,13 @@ import org.jetbrains.annotations.Range;
 public class Material {
 
     private @NotNull final ItemType itemType;
-    private @Range(from = 0, to = Integer.MAX_VALUE) final byte data;
+    private @Range(from = 0, to = Byte.MAX_VALUE) final byte data;
 
     public Material(@NotNull ItemType itemType) {
         this((byte) 0, itemType);
     }
 
-    public Material(@Range(from = 0, to = Integer.MAX_VALUE) byte data, @NotNull ItemType itemType) {
+    public Material(@Range(from = 0, to = Byte.MAX_VALUE) byte data, @NotNull ItemType itemType) {
         this.itemType = itemType;
         this.data = data;
     }
